[#11] Color Design Token 추가
